Time: 27 ms (14.06%), Space: 84.5 MB (99.2%) - LeetHub

+++ 3634-minimum-removals-to-balance-array/3634-minimum-removals-to-balance-array.java
@@ -2,11 +2,12 @@ class Solution {
     public int minRemoval(int[] nums, int k) {
         int m = nums.length;
         Arrays.sort(nums);
+
         int i = 0;
         int len = 0;
 
         for(int j = 0 ; j < m ; j++){
-            while((long) nums[j] > (long) nums[i] * k){
+            while(nums[j] > (long) nums[i] * k){
                 i++;
             }
             len = Math.max(len , j - i + 1);
